+++ lib/src/renderer/interfaces/platform.dart
@@ -15,9 +15,11 @@ abstract class PdfxPlatform {
   /// The instance of [PdfxPlatform] to use.
   static final PdfxPlatform instance = PdfxPlatformPigeon();
 
-  Future<PdfDocument> openFile(String filePath);
+  Future<bool> isPasswordSupported();
 
-  Future<PdfDocument> openAsset(String name);
+  Future<PdfDocument> openFile(String filePath, {String? password});
 
-  Future<PdfDocument> openData(FutureOr<Uint8List> data);
+  Future<PdfDocument> openAsset(String name, {String? password});
+
+  Future<PdfDocument> openData(FutureOr<Uint8List> data, {String? password});
 }